Problem 8 fix and gitignore
Fixed problem 8, updated .gitignore so it wouldn't include compiled
files by C++

Will leave in .cbp files for other code::blocks users.

+++ Problem Solver/Problem8.h
@@ -4,6 +4,6 @@ class Problem8
 public:
 	Problem8();
 	virtual ~Problem8();
-	long long Solve();
+	int Solve();
 };
